Fix Gen IV met locations

+++ Pkmds.Web/Services/IAppService.cs
@@ -34,7 +34,7 @@ public interface IAppService
 
     IEnumerable<ComboItem> SearchMetLocations(string searchString, bool isEggLocation = false);
 
-    ComboItem GetMetLocationComboItem(int metLocationId, bool isEggLocation = false);
+    ComboItem GetMetLocationComboItem(ushort metLocationId, bool isEggLocation = false);
 
     IEnumerable<ComboItem> SearchMoves(string searchString);
